fixed many to many relation of MOVIE-ACTOR

+++ DSS4-MS-MOVIE-V1/src/main/java/com/dss/movie/service/MovieServiceImpl.java
@@ -1,17 +1,39 @@
 package com.dss.movie.service;
 
+import com.dss.movie.exception.AbstractRuntimeException;
+import com.dss.movie.exception.ActorNotFoundException;
+import com.dss.movie.exception.InvalidMovieException;
+import com.dss.movie.exception.MovieNotFoundException;
+import com.dss.movie.model.Actor;
 import com.dss.movie.model.Movie;
+import com.dss.movie.model.MovieDto;
+import com.dss.movie.proxy.ActorProxy;
 import com.dss.movie.repository.MovieDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
+import java.time.LocalDate;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 public class MovieServiceImpl implements MovieService {
 
     @Autowired MovieDao movieDao;
+    @Autowired ActorProxy actorProxy;
+
+    private void validateActorList(Movie movie) throws AbstractRuntimeException  {
+        if(movie.getActorList().size() > 0){
+            Map<String, Actor> actorMap = actorProxy.findAll().stream()
+                    .collect(Collectors.toMap(Actor::getId, actor -> actor));
+
+            for(Actor actor: movie.getActorList()){
+                if(actorMap.put(actor.getId(), actor) == null){
+                    throw new ActorNotFoundException("Actor ID does not exist");
+                }
+            }
+        }
+    }
 
     @Override
     public Movie findById(String id) {
@@ -29,22 +51,25 @@ public class MovieServiceImpl implements MovieService {
 
     @Override
     public String save(Movie movie) {
-        movie = movieDao.save(movie);
-        return movie.getId();
+        if(movie.getActorList() != null){
+            validateActorList(movie);
+        }
+        return movieDao.save(movie).getId();
     }
 
     @Override
-    public boolean update(Movie movie) {
+    public boolean update(String id, MovieDto newModel) {
         Boolean isSuccess = true;
-        try {
-            if(movie.getId()==null){
-                throw new Exception();
-            } else {
-                movieDao.save(movie);
-            }
-        } catch(Exception e){
+
+        Optional<Movie> temp = movieDao.findById(id);
+        if(temp.isPresent()){
+            Movie oldModel = temp.get();
+            oldModel.setProductionCost(newModel.getProductionCost()!=null ? newModel.getProductionCost() : oldModel.getProductionCost());
+            oldModel.setImageDirectory(newModel.getImageDirectory()!=null ? newModel.getImageDirectory() : oldModel.getImageDirectory());
+            movieDao.save(oldModel);
+        } else {
             isSuccess = false;
-            e.printStackTrace();
+            throw new MovieNotFoundException("Movie ID does not exist");
         }
         return isSuccess;
     }
@@ -52,10 +77,19 @@ public class MovieServiceImpl implements MovieService {
     @Override
     public boolean deleteById(String id) {
         Boolean isSuccess = true;
-        try {
-            movieDao.deleteById(id);
-        } catch (Exception e){
+
+        Optional<Movie> movie = movieDao.findById(id);
+        if(movie.isPresent()){
+            int dateNow = LocalDate.now().getYear();
+            int datePast = Integer.parseInt(movie.get().getYearReleased());
+            if(dateNow - datePast > 1){
+                movieDao.deleteById(id);
+            } else {
+                throw new InvalidMovieException("Unable to delete movie released < 1 year");
+            }
+        } else {
             isSuccess = false;
+            throw new MovieNotFoundException("Movie ID does not exist");
         }
         return isSuccess;
     }
